e2e: move std and manchette const on a seprate file
